+++ src/Curiosity.Migrations.SqlServer/MigrationActionHelper.cs
@@ -28,17 +28,19 @@ internal class MigrationActionHelper
     /// <param name="func">Func to execute.</param>
     /// <param name="errorCode">Type of error if occurred.</param>
     /// <param name="errorMessage">Error message if error occurred.</param>
+    /// <param name="cancellationToken"></param>
     /// <typeparam name="T">Result type of func.</typeparam>
     /// <returns>Result of <see cref="func"/>.</returns>
     /// <exception cref="MigrationException">If <see cref="func"/> throw exception.</exception>
     internal async Task<T> TryExecuteAsync<T>(
-        Func<Task<T>> func,
+        Func<CancellationToken, Task<T>> func,
         MigrationErrorCode errorCode,
-        string errorMessage)
+        string errorMessage,
+        CancellationToken cancellationToken = default)
     {
         try
         {
-            return await func();
+            return await func(cancellationToken);
         }
         catch (MigrationException)
         {
@@ -52,42 +54,9 @@ internal class MigrationActionHelper
         {
             throw new MigrationException(
                 errorCode,
-                $"{errorMessage} Database: {_databaseName}",
-                e);
-        }
-    }
-
-    /// <summary>
-    /// Tries to execute async function with error handling.
-    /// </summary>
-    /// <param name="func">Func to execute.</param>
-    /// <param name="errorCode">Type of error if occurred.</param>
-    /// <param name="errorMessage">Error message if error occurred.</param>
-    /// <returns>Result of <see cref="func"/>.</returns>
-    /// <exception cref="MigrationException">If <see cref="func"/> throw exception.</exception>
-    internal async Task TryExecuteAsync(
-        Func<Task> func,
-        MigrationErrorCode errorCode,
-        string errorMessage)
-    {
-        try
-        {
-            await func();
-        }
-        catch (MigrationException)
-        {
-            throw;
-        }
-        catch (SqlException sqlEx)
-        {
-            throw CreateMigrationExceptionFromSqlException(sqlEx, errorCode, errorMessage);
-        }
-        catch (Exception e)
-        {
-            throw new MigrationException(
-                errorCode,
-                $"{errorMessage} Database: {_databaseName}",
-                e);
+                errorMessage,
+                e,
+                _databaseName);
         }
     }
 
@@ -122,8 +91,9 @@ internal class MigrationActionHelper
         {
             throw new MigrationException(
                 errorCode,
-                $"{errorMessage} Database: {_databaseName}",
-                e);
+                errorMessage,
+                e,
+                _databaseName);
         }
     }
 
@@ -139,12 +109,13 @@ internal class MigrationActionHelper
         var errorCode = GetErrorCodeFromSqlException(sqlEx, defaultErrorCode);
         
         // Add SQL error information to the error message
-        var detailedErrorMessage = $"{errorMessage}. SQL Error: {sqlEx.Number}, State: {sqlEx.State}, Message: {sqlEx.Message}. Database: {_databaseName}";
+        var detailedErrorMessage = $"{errorMessage}. SQL Error: {sqlEx.Number}, State: {sqlEx.State}, Message: {sqlEx.Message}";
         
         return new MigrationException(
             errorCode,
             detailedErrorMessage,
-            sqlEx);
+            sqlEx,
+            _databaseName);
     }
 
     /// <summary>