+++ tests/IntegrationTests/Curiosity.Migrations.PostgreSql.IntegrationTests/Fixtures/PostgresContainerFixture.cs
@@ -31,7 +31,7 @@ public class PostgresContainerFixture : IAsyncLifetime
     public PostgresContainerFixture()
     {
         Container = new PostgreSqlBuilder()
-            .WithImage("postgres:latest")
+            .WithImage("postgres:14.17-bookworm")
             .WithUsername(Username)
             .WithPassword(Password)
             .WithDatabase("postgres")