+++ tests/IntegrationTests/Curiosity.Migrations.SqlServer.IntegrationTests/SqlServerDbCreationTests.cs
@@ -175,7 +175,6 @@ public class SqlServerDbCreationTests : IClassFixture<SqlServerContainerFixture>
             var options = new SqlServerMigrationConnectionOptions(
                 _containerFixture.GetConnectionString(dbName),
                 collation: "SQL_Latin1_General_CP1_CI_AS",
-                maxConnections: 10,
                 initialSize: 10,
                 maxSize: 100,
                 fileGrowth: 5,